backend/python/research-agent/requirements-container.txt
bedrock-agentcore<=0.1.5
strands-agents[bedrock]>=1.36.0
boto3>=1.34.0
